+++ src/components/features/search/tabs/gates-tab/gates-tab.tsx
