+++ src/components/features/search/tabs/gates-tab/index.ts
