+++ src/components/features/search/tabs/gates-tab/table-view.tsx
