+++ src/components/features/search/tabs/people-tab/index.ts
