+++ src/components/features/search/tabs/people-tab/people-tab.tsx
