+++ src/components/features/search/tabs/people-tab/table-view.tsx
